refactor(block): 重构方块物品类并添加新功能
- 将方块物品类从 item 包移动到 block/item 包
- 添加 MengerSpongeBlockItem 新类，实现与炼药锅的交互
- 修改 ModBlocks 文件，使用新路径引用方块物品类
- 更新 ModDispenserBehavior，添加门格海绵 的特殊行为
