+++ src/main/java/dev/dubhe/anvilcraft/block/item/ChuteBlockItem.java
@@ -1,4 +1,4 @@
-package dev.dubhe.anvilcraft.item;
+package dev.dubhe.anvilcraft.block.item;
 
 import dev.dubhe.anvilcraft.api.itemhandler.IItemHandlerHolder;
 import net.minecraft.MethodsReturnNonnullByDefault;
