+++ src/main/java/dev/dubhe/anvilcraft/block/item/FlexibleMultiPartBlockItem.java
@@ -1,7 +1,7 @@
-package dev.dubhe.anvilcraft.item;
+package dev.dubhe.anvilcraft.block.item;
 
-import dev.dubhe.anvilcraft.block.multipart.SimpleMultiPartBlock;
-import dev.dubhe.anvilcraft.block.state.ISimpleMultiPartBlockState;
+import dev.dubhe.anvilcraft.block.multipart.FlexibleMultiPartBlock;
+import dev.dubhe.anvilcraft.block.state.IFlexibleMultiPartBlockState;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -13,16 +13,17 @@ import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.block.state.properties.Property;
 import net.minecraft.world.phys.BlockHitResult;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 
 @ParametersAreNonnullByDefault
 @MethodsReturnNonnullByDefault
-public class SimpleMultiPartBlockItem<P extends Enum<P> & ISimpleMultiPartBlockState<P>> extends BlockItem {
-    private final SimpleMultiPartBlock<P> block;
+public class FlexibleMultiPartBlockItem<P extends Enum<P> & IFlexibleMultiPartBlockState<P, E>, T extends Property<E>, E extends Comparable<E>> extends BlockItem {
+    private final FlexibleMultiPartBlock<P, T, E> block;
 
-    public SimpleMultiPartBlockItem(SimpleMultiPartBlock<P> block, Properties properties) {
+    public FlexibleMultiPartBlockItem(FlexibleMultiPartBlock<P, T, E> block, Properties properties) {
         super(block, properties);
         this.block = block;
     }
@@ -32,7 +33,7 @@ public class SimpleMultiPartBlockItem<P extends Enum<P> & ISimpleMultiPartBlockS
         BlockPos pos = context.getClickedPos();
         Level level = context.getLevel();
         for (P part : this.block.getParts()) {
-            BlockPos offset = pos.offset(part.getOffset());
+            BlockPos offset = pos.offset(part.getOffset(state.getValue(block.getAdditionalProperty())));
             BlockState blockState =
                 level.isWaterAt(offset) ? Blocks.WATER.defaultBlockState() : Blocks.AIR.defaultBlockState();
             level.setBlock(offset, blockState, 27);
@@ -40,13 +41,13 @@ public class SimpleMultiPartBlockItem<P extends Enum<P> & ISimpleMultiPartBlockS
         return super.placeBlock(context, state);
     }
 
-    private int getMaxOffsetDistance(Direction clickedFace) {
+    private int getMaxOffsetDistance(BlockState state, Direction clickedFace) {
         Vec3i normal = clickedFace.getOpposite().getNormal();
         int i = 0;
         for (P part : this.block.getParts()) {
-            int x = part.getOffsetX() * normal.getX()
-                + part.getOffsetY() * normal.getY()
-                + part.getOffsetZ() * normal.getZ();
+            int x = part.getOffsetX(state.getValue(block.getAdditionalProperty())) * normal.getX()
+                + part.getOffsetY(state.getValue(block.getAdditionalProperty())) * normal.getY()
+                + part.getOffsetZ(state.getValue(block.getAdditionalProperty())) * normal.getZ();
             i = Math.max(x, i);
         }
         return ++i;
@@ -56,6 +57,8 @@ public class SimpleMultiPartBlockItem<P extends Enum<P> & ISimpleMultiPartBlockS
     public InteractionResult useOn(UseOnContext context) {
         InteractionResult result = super.useOn(context);
         Direction clickedFace = context.getClickedFace();
+        BlockState state = this.block.getPlacementState(new BlockPlaceContext(context));
+        if (state == null) return InteractionResult.FAIL;
         if (result == InteractionResult.FAIL) {
             return super.useOn(new UseOnContext(
                 context.getLevel(),
@@ -63,11 +66,17 @@ public class SimpleMultiPartBlockItem<P extends Enum<P> & ISimpleMultiPartBlockS
                 context.getHand(),
                 context.getItemInHand(),
                 new BlockHitResult(
-                    context.getClickLocation().relative(clickedFace, this.getMaxOffsetDistance(clickedFace)),
+                    context.getClickLocation().relative(clickedFace, this.getMaxOffsetDistance(state, clickedFace)),
                     clickedFace,
-                    context.getClickedPos().relative(clickedFace, this.getMaxOffsetDistance(clickedFace)),
+                    context.getClickedPos().relative(clickedFace, this.getMaxOffsetDistance(state, clickedFace)),
                     false)));
         }
         return result;
+    }
+
+    @Override
+    public boolean canPlace(BlockPlaceContext context, BlockState state) {
+        if (!block.hasEnoughSpace(state, context.getClickedPos(), context.getLevel())) return false;
+        return super.canPlace(context, state);
     }
 }
