+++ src/main/java/dev/dubhe/anvilcraft/block/item/MengerSpongeBlockItem.java
@@ -1,0 +1,28 @@
+package dev.dubhe.anvilcraft.block.item;
+
+import net.minecraft.core.BlockPos;
+import net.minecraft.world.InteractionResult;
+import net.minecraft.world.item.BlockItem;
+import net.minecraft.world.item.context.UseOnContext;
+import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.AbstractCauldronBlock;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
+
+public class MengerSpongeBlockItem extends BlockItem {
+    public MengerSpongeBlockItem(Block block, Properties properties) {
+        super(block, properties);
+    }
+
+    @Override
+    public InteractionResult useOn(UseOnContext context) {
+        Level level = context.getLevel();
+        BlockPos pos = context.getClickedPos();
+        BlockState state = level.getBlockState(pos);
+        if (state.getBlock() instanceof AbstractCauldronBlock) {
+            level.setBlockAndUpdate(pos, Blocks.CAULDRON.defaultBlockState());
+        }
+        return InteractionResult.SUCCESS;
+    }
+}
