+++ src/main/java/dev/dubhe/anvilcraft/block/item/SuperHeavyBlockItem.java
@@ -1,24 +1,20 @@
-package dev.dubhe.anvilcraft.item.abnormal;
+package dev.dubhe.anvilcraft.block.item;
 
-import net.minecraft.MethodsReturnNonnullByDefault;
+import dev.dubhe.anvilcraft.item.abnormal.ISuperHeavy;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 
-import javax.annotation.ParametersAreNonnullByDefault;
-
-@ParametersAreNonnullByDefault
-@MethodsReturnNonnullByDefault
-public class CursedBlockItem extends BlockItem implements ICursed {
-    public CursedBlockItem(Block block, Properties properties) {
+public class SuperHeavyBlockItem extends BlockItem implements ISuperHeavy {
+    public SuperHeavyBlockItem(Block block, Properties properties) {
         super(block, properties);
     }
 
     @Override
     public void inventoryTick(ItemStack stack, Level level, Entity entity, int slotId, boolean isSelected) {
         super.inventoryTick(stack, level, entity, slotId, isSelected);
-        ICursed.super.inventoryTick(stack, level, entity, slotId, isSelected);
+        ISuperHeavy.super.inventoryTick(stack, level, entity, slotId, isSelected);
     }
 }
