+++ src/main/java/dev/dubhe/anvilcraft/block/item/package-info.java
@@ -1,0 +1,7 @@
+@ParametersAreNonnullByDefault
+@MethodsReturnNonnullByDefault
+package dev.dubhe.anvilcraft.block.item;
+
+import com.mojang.math.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;
