+++ src/main/java/dev/dubhe/anvilcraft/item/TeslaTowerItem.java
@@ -1,6 +1,7 @@
 package dev.dubhe.anvilcraft.item;
 
 import dev.dubhe.anvilcraft.block.entity.TeslaTowerBlockEntity;
+import dev.dubhe.anvilcraft.block.item.SimpleMultiPartBlockItem;
 import dev.dubhe.anvilcraft.block.multipart.SimpleMultiPartBlock;
 import dev.dubhe.anvilcraft.block.state.Vertical4PartHalf;
 import net.minecraft.core.BlockPos;
@@ -8,7 +9,6 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
-import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
 import javax.annotation.ParametersAreNonnullByDefault;
@@ -21,11 +21,11 @@ public class TeslaTowerItem extends SimpleMultiPartBlockItem<Vertical4PartHalf> 
 
     @Override
     protected boolean updateCustomBlockEntityTag(
-        @NotNull BlockPos pos,
-        @NotNull Level level,
+        BlockPos pos,
+        Level level,
         @Nullable Player player,
-        @NotNull ItemStack stack,
-        @NotNull BlockState state) {
+        ItemStack stack,
+        BlockState state) {
         if (!(level.getBlockEntity(pos) instanceof TeslaTowerBlockEntity teslaTowerBlockEntity)) return false;
         if (player == null) return false;
         teslaTowerBlockEntity.initWhiteList(player);
